Reworked command module, moved heavy work to lib module.
