+++ src/command/init.rs
@@ -1,0 +1,10 @@
+use crate::lib;
+
+pub fn new() -> clap::Command {
+    clap::Command::new("init")
+        .about("Creates a new C/C++ project within the current directory.")
+}
+
+pub fn exec() {
+    lib::new_project::initialize_project(None);
+}
