+++ src/command/mod.rs
@@ -1,4 +1,5 @@
-pub mod build_project;
+pub mod build;
 pub mod configure;
+pub mod init;
 pub mod install;
-pub mod new_project;
+pub mod new;
