+++ src/command/new.rs
@@ -1,0 +1,15 @@
+use crate::lib;
+
+pub fn new() -> clap::Command {
+    clap::Command::new("new")
+        .about("Creates a new C/C++ project.")
+        .arg(
+            clap::Arg::new("folder")
+                .required(true)
+                .help("The folder to create for the new project."),
+        )
+}
+
+pub fn exec(arg_matches: &clap::ArgMatches) {
+    lib::new_project::initialize_project(arg_matches.get_one::<String>("folder"));
+}
