+++ src/lib/mod.rs
@@ -1,0 +1,4 @@
+pub mod build_project;
+pub mod configure;
+pub mod install;
+pub mod new_project;
