+++ src/main.rs
@@ -1,15 +1,11 @@
-use crate::command::build_project::build_project;
-use crate::command::configure::configure;
-use crate::command::install::install;
-use crate::command::new_project::initialize_project;
-
-use clap::{crate_authors, crate_description, crate_name, crate_version, Arg, ArgAction, Command};
+use clap::{crate_authors, crate_description, crate_name, crate_version, Command};
 
 #[cfg(windows)]
 use fansi::windows::enable_ansi_support;
 
 mod command;
 mod config;
+mod lib;
 mod settings;
 mod terminal;
 
@@ -21,52 +17,11 @@ fn main() {
         .author(crate_authors!("\n"))
         .about(crate_description!())
         .version(crate_version!())
-        .subcommand(
-            Command::new("new")
-                .about("Creates a new C/C++ project.")
-                .arg(
-                    Arg::new("folder")
-                        .required(true)
-                        .help("The folder to create for the new project."),
-                ),
-        )
-        .subcommand(
-            Command::new("init").about("Creates a new C/C++ project within the current directory."),
-        )
-        .subcommand(
-            Command::new("build")
-                .about("Builds the existing C/C++ project within the current directory.")
-                .arg(
-                    Arg::new("static-library")
-                        .help("Specifies that a static library should be built from the current project.")
-                        .short('s')
-                        .long("static")
-                        .action(ArgAction::SetTrue),
-                )
-        )
-        .subcommand(
-            Command::new("install")
-                .about("Installs a tool or library.")
-                .arg(
-                    Arg::new("tool-library-name")
-                        .required(true)
-                        .help("The name of the tool or library to install."),
-                ),
-        )
-        .subcommand(
-            Command::new("configure")
-                .about("Configure CCake's global properties.")
-                .arg(
-                    Arg::new("default-c-compiler-dir")
-                        .help("The path to the default C compiler to be used for building projects.")
-                        .long("c-compiler-dir"),
-                )
-                .arg(
-                    Arg::new("default-cpp-compiler-dir")
-                        .help("The path to the default C++ compiler to be used for building projects.")
-                        .long("cpp-compiler-dir"),
-                ),
-        );
+        .subcommand(command::new::new())
+        .subcommand(command::init::new())
+        .subcommand(command::build::new())
+        .subcommand(command::install::new())
+        .subcommand(command::configure::new());
 
     let matches = command.get_matches();
 
@@ -74,13 +29,11 @@ fn main() {
     enable_ansi_support().expect("Failed to enable ansi support for Windows!"); // TODO: Remove this line.
 
     match matches.subcommand() {
-        Some(("new", arg_matches)) => initialize_project(arg_matches.get_one::<String>("folder")),
-        Some(("init", _)) => initialize_project(None),
-        Some(("build", arg_matches)) => build_project(arg_matches),
-        Some(("install", arg_matches)) => {
-            install(arg_matches.get_one::<String>("tool-library-name"))
-        }
-        Some(("configure", arg_matches)) => configure(arg_matches),
+        Some(("new", arg_matches)) => command::new::exec(arg_matches),
+        Some(("init", _)) => command::init::exec(),
+        Some(("build", arg_matches)) => command::build::exec(arg_matches),
+        Some(("install", arg_matches)) => command::install::exec(arg_matches),
+        Some(("configure", arg_matches)) => command::configure::exec(arg_matches),
         Some((_, _)) => (),
         None => (),
     }
